Move logic from annotateBinaryOperator into BinaryOperator
