+++ src/expressions/operators/arithmetic/BinaryOperator.ts
@@ -1,29 +1,31 @@
-import { BinaryEvaluationFunction } from '../../../typeInference/binaryEvaluationFunction';
-import atomize from '../../dataTypes/atomize';
-import castToType from '../../dataTypes/castToType';
-import createAtomicValue from '../../dataTypes/createAtomicValue';
-import isSubtypeOf from '../../dataTypes/isSubtypeOf';
-import sequenceFactory from '../../dataTypes/sequenceFactory';
-import { SequenceType, ValueType } from '../../dataTypes/Value';
+import AtomicValue from '../../../expressions/dataTypes/AtomicValue';
+import castToType from '../../../expressions/dataTypes/castToType';
+import createAtomicValue from '../../../expressions/dataTypes/createAtomicValue';
+import isSubtypeOf from '../../../expressions/dataTypes/isSubtypeOf';
+import { ValueType } from '../../../expressions/dataTypes/Value';
 import {
 	addDuration as addDurationToDateTime,
 	subtract as dateTimeSubtract,
 	subtractDuration as subtractDurationFromDateTime,
-} from '../../dataTypes/valueTypes/DateTime';
+} from '../../../expressions/dataTypes/valueTypes/DateTime';
 import {
 	add as dayTimeDurationAdd,
 	divide as dayTimeDurationDivide,
 	divideByDayTimeDuration as dayTimeDurationDivideByDayTimeDuration,
 	multiply as dayTimeDurationMultiply,
 	subtract as dayTimeDurationSubtract,
-} from '../../dataTypes/valueTypes/DayTimeDuration';
+} from '../../../expressions/dataTypes/valueTypes/DayTimeDuration';
 import {
 	add as yearMonthDurationAdd,
 	divide as yearMonthDurationDivide,
 	divideByYearMonthDuration as yearMonthDurationDivideByYearMonthDuration,
 	multiply as yearMonthDurationMultiply,
 	subtract as yearMonthDurationSubtract,
-} from '../../dataTypes/valueTypes/YearMonthDuration';
+} from '../../../expressions/dataTypes/valueTypes/YearMonthDuration';
+import { BinaryEvaluationFunction } from '../../../typeInference/binaryEvaluationFunction';
+import atomize from '../../dataTypes/atomize';
+import sequenceFactory from '../../dataTypes/sequenceFactory';
+import { SequenceType } from '../../dataTypes/Value';
 import Expression from '../../Expression';
 
 function determineReturnType(typeA: ValueType, typeB: ValueType): ValueType {
@@ -40,11 +42,10 @@ function determineReturnType(typeA: ValueType, typeB: ValueType): ValueType {
 }
 
 function generateBinaryOperatorFunction(
-	operator,
+	operator: string,
 	typeA: ValueType,
-	typeB: ValueType,
-	retType?: SequenceType
-) {
+	typeB: ValueType
+): [BinaryEvaluationFunction, ValueType] {
 	let castFunctionForValueA = null;
 	let castFunctionForValueB = null;
 
@@ -57,7 +58,7 @@ function generateBinaryOperatorFunction(
 		typeB = ValueType.XSDOUBLE;
 	}
 
-	function applyCastFunctions(valueA, valueB) {
+	function applyCastFunctions(valueA: AtomicValue, valueB: AtomicValue) {
 		return {
 			castA: castFunctionForValueA ? castFunctionForValueA(valueA) : valueA,
 			castB: castFunctionForValueB ? castFunctionForValueB(valueB) : valueB,
@@ -68,66 +69,84 @@ function generateBinaryOperatorFunction(
 		switch (operator) {
 			case 'addOp': {
 				const returnType = determineReturnType(typeA, typeB);
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
 
-					return createAtomicValue(castA.value + castB.value, returnType);
-				};
+						return createAtomicValue(castA.value + castB.value, returnType);
+					},
+					returnType,
+				];
 			}
 			case 'subtractOp': {
 				const returnType = determineReturnType(typeA, typeB);
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(castA.value - castB.value, returnType);
-				};
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
+						return createAtomicValue(castA.value - castB.value, returnType);
+					},
+					returnType,
+				];
 			}
 			case 'multiplyOp': {
 				const returnType = determineReturnType(typeA, typeB);
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(castA.value * castB.value, returnType);
-				};
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
+						return createAtomicValue(castA.value * castB.value, returnType);
+					},
+					returnType,
+				];
 			}
 			case 'divOp': {
 				let returnType = determineReturnType(typeA, typeB);
 				if (returnType === ValueType.XSINTEGER) {
 					returnType = ValueType.XSDECIMAL;
 				}
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(castA.value / castB.value, returnType);
-				};
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
+						return createAtomicValue(castA.value / castB.value, returnType);
+					},
+					returnType,
+				];
 			}
 			case 'idivOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					if (castB.value === 0) {
-						throw new Error('FOAR0001: Divisor of idiv operator cannot be (-)0');
-					}
-					if (
-						Number.isNaN(castA.value) ||
-						Number.isNaN(castB.value) ||
-						!Number.isFinite(castA.value)
-					) {
-						throw new Error(
-							'FOAR0002: One of the operands of idiv is NaN or the first operand is (-)INF'
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
+						if (castB.value === 0) {
+							throw new Error('FOAR0001: Divisor of idiv operator cannot be (-)0');
+						}
+						if (
+							Number.isNaN(castA.value) ||
+							Number.isNaN(castB.value) ||
+							!Number.isFinite(castA.value)
+						) {
+							throw new Error(
+								'FOAR0002: One of the operands of idiv is NaN or the first operand is (-)INF'
+							);
+						}
+						if (Number.isFinite(castA.value) && !Number.isFinite(castB.value)) {
+							return createAtomicValue(0, ValueType.XSINTEGER);
+						}
+						return createAtomicValue(
+							Math.trunc(castA.value / castB.value),
+							ValueType.XSINTEGER
 						);
-					}
-					if (Number.isFinite(castA.value) && !Number.isFinite(castB.value)) {
-						return createAtomicValue(0, ValueType.XSINTEGER);
-					}
-					return createAtomicValue(
-						Math.trunc(castA.value / castB.value),
-						ValueType.XSINTEGER
-					);
-				};
+					},
+					ValueType.XSINTEGER,
+				];
 			case 'modOp': {
 				const returnType = determineReturnType(typeA, typeB);
 
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(castA.value % castB.value, returnType);
-				};
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
+						return createAtomicValue(castA.value % castB.value, returnType);
+					},
+					returnType,
+				];
 			}
 		}
 	}
@@ -138,29 +157,38 @@ function generateBinaryOperatorFunction(
 	) {
 		switch (operator) {
 			case 'addOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						yearMonthDurationAdd(castA.value, castB.value),
-						ValueType.XSYEARMONTHDURATION
-					);
-				};
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
+						return createAtomicValue(
+							yearMonthDurationAdd(castA.value, castB.value),
+							ValueType.XSYEARMONTHDURATION
+						);
+					},
+					ValueType.XSYEARMONTHDURATION,
+				];
 			case 'subtractOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						yearMonthDurationSubtract(castA.value, castB.value),
-						ValueType.XSYEARMONTHDURATION
-					);
-				};
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
+						return createAtomicValue(
+							yearMonthDurationSubtract(castA.value, castB.value),
+							ValueType.XSYEARMONTHDURATION
+						);
+					},
+					ValueType.XSYEARMONTHDURATION,
+				];
 			case 'divOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						yearMonthDurationDivideByYearMonthDuration(castA.value, castB.value),
-						ValueType.XSDECIMAL
-					);
-				};
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
+						return createAtomicValue(
+							yearMonthDurationDivideByYearMonthDuration(castA.value, castB.value),
+							ValueType.XSDECIMAL
+						);
+					},
+					ValueType.XSDECIMAL,
+				];
 		}
 	}
 
@@ -170,21 +198,27 @@ function generateBinaryOperatorFunction(
 	) {
 		switch (operator) {
 			case 'multiplyOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						yearMonthDurationMultiply(castA.value, castB.value),
-						ValueType.XSYEARMONTHDURATION
-					);
-				};
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
+						return createAtomicValue(
+							yearMonthDurationMultiply(castA.value, castB.value),
+							ValueType.XSYEARMONTHDURATION
+						);
+					},
+					ValueType.XSYEARMONTHDURATION,
+				];
 			case 'divOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						yearMonthDurationDivide(castA.value, castB.value),
-						ValueType.XSYEARMONTHDURATION
-					);
-				};
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
+						return createAtomicValue(
+							yearMonthDurationDivide(castA.value, castB.value),
+							ValueType.XSYEARMONTHDURATION
+						);
+					},
+					ValueType.XSYEARMONTHDURATION,
+				];
 		}
 	}
 
@@ -193,13 +227,16 @@ function generateBinaryOperatorFunction(
 		isSubtypeOf(typeB, ValueType.XSYEARMONTHDURATION)
 	) {
 		if (operator === 'multiplyOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					yearMonthDurationMultiply(castB.value, castA.value),
-					ValueType.XSYEARMONTHDURATION
-				);
-			};
+			return [
+				(a, b) => {
+					const { castA, castB } = applyCastFunctions(a, b);
+					return createAtomicValue(
+						yearMonthDurationMultiply(castB.value, castA.value),
+						ValueType.XSYEARMONTHDURATION
+					);
+				},
+				ValueType.XSYEARMONTHDURATION,
+			];
 		}
 	}
 
@@ -209,29 +246,38 @@ function generateBinaryOperatorFunction(
 	) {
 		switch (operator) {
 			case 'addOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						dayTimeDurationAdd(castA.value, castB.value),
-						ValueType.XSDAYTIMEDURATION
-					);
-				};
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
+						return createAtomicValue(
+							dayTimeDurationAdd(castA.value, castB.value),
+							ValueType.XSDAYTIMEDURATION
+						);
+					},
+					ValueType.XSDAYTIMEDURATION,
+				];
 			case 'subtractOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						dayTimeDurationSubtract(castA.value, castB.value),
-						ValueType.XSDAYTIMEDURATION
-					);
-				};
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
+						return createAtomicValue(
+							dayTimeDurationSubtract(castA.value, castB.value),
+							ValueType.XSDAYTIMEDURATION
+						);
+					},
+					ValueType.XSDAYTIMEDURATION,
+				];
 			case 'divOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						dayTimeDurationDivideByDayTimeDuration(castA.value, castB.value),
-						ValueType.XSDECIMAL
-					);
-				};
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
+						return createAtomicValue(
+							dayTimeDurationDivideByDayTimeDuration(castA.value, castB.value),
+							ValueType.XSDECIMAL
+						);
+					},
+					ValueType.XSDECIMAL,
+				];
 		}
 	}
 	if (
@@ -240,21 +286,27 @@ function generateBinaryOperatorFunction(
 	) {
 		switch (operator) {
 			case 'multiplyOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						dayTimeDurationMultiply(castA.value, castB.value),
-						ValueType.XSDAYTIMEDURATION
-					);
-				};
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
+						return createAtomicValue(
+							dayTimeDurationMultiply(castA.value, castB.value),
+							ValueType.XSDAYTIMEDURATION
+						);
+					},
+					ValueType.XSDAYTIMEDURATION,
+				];
 			case 'divOp':
-				return (a, b) => {
-					const { castA, castB } = applyCastFunctions(a, b);
-					return createAtomicValue(
-						dayTimeDurationDivide(castA.value, castB.value),
-						ValueType.XSDAYTIMEDURATION
-					);
-				};
+				return [
+					(a, b) => {
+						const { castA, castB } = applyCastFunctions(a, b);
+						return createAtomicValue(
+							dayTimeDurationDivide(castA.value, castB.value),
+							ValueType.XSDAYTIMEDURATION
+						);
+					},
+					ValueType.XSDAYTIMEDURATION,
+				];
 		}
 	}
 	if (
@@ -262,13 +314,16 @@ function generateBinaryOperatorFunction(
 		isSubtypeOf(typeB, ValueType.XSDAYTIMEDURATION)
 	) {
 		if (operator === 'multiplyOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					dayTimeDurationMultiply(castB.value, castA.value),
-					ValueType.XSDAYTIMEDURATION
-				);
-			};
+			return [
+				(a, b) => {
+					const { castA, castB } = applyCastFunctions(a, b);
+					return createAtomicValue(
+						dayTimeDurationMultiply(castB.value, castA.value),
+						ValueType.XSDAYTIMEDURATION
+					);
+				},
+				ValueType.XSDAYTIMEDURATION,
+			];
 		}
 	}
 
@@ -278,13 +333,16 @@ function generateBinaryOperatorFunction(
 		(isSubtypeOf(typeA, ValueType.XSTIME) && isSubtypeOf(typeB, ValueType.XSTIME))
 	) {
 		if (operator === 'subtractOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					dateTimeSubtract(castA.value, castB.value),
-					ValueType.XSDAYTIMEDURATION
-				);
-			};
+			return [
+				(a, b) => {
+					const { castA, castB } = applyCastFunctions(a, b);
+					return createAtomicValue(
+						dateTimeSubtract(castA.value, castB.value),
+						ValueType.XSDAYTIMEDURATION
+					);
+				},
+				ValueType.XSDAYTIMEDURATION,
+			];
 		}
 
 		throw new Error(`XPTY0004: ${operator} not available for types ${typeA} and ${typeB}`);
@@ -297,22 +355,28 @@ function generateBinaryOperatorFunction(
 			isSubtypeOf(typeB, ValueType.XSDAYTIMEDURATION))
 	) {
 		if (operator === 'addOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					addDurationToDateTime(castA.value, castB.value),
-					ValueType.XSDATETIME
-				);
-			};
+			return [
+				(a, b) => {
+					const { castA, castB } = applyCastFunctions(a, b);
+					return createAtomicValue(
+						addDurationToDateTime(castA.value, castB.value),
+						ValueType.XSDATETIME
+					);
+				},
+				ValueType.XSDATETIME,
+			];
 		}
 		if (operator === 'subtractOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					subtractDurationFromDateTime(castA.value, castB.value),
-					ValueType.XSDATETIME
-				);
-			};
+			return [
+				(a, b) => {
+					const { castA, castB } = applyCastFunctions(a, b);
+					return createAtomicValue(
+						subtractDurationFromDateTime(castA.value, castB.value),
+						ValueType.XSDATETIME
+					);
+				},
+				ValueType.XSDATETIME,
+			];
 		}
 	}
 
@@ -322,43 +386,55 @@ function generateBinaryOperatorFunction(
 		(isSubtypeOf(typeA, ValueType.XSDATE) && isSubtypeOf(typeB, ValueType.XSDAYTIMEDURATION))
 	) {
 		if (operator === 'addOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					addDurationToDateTime(castA.value, castB.value),
-					ValueType.XSDATE
-				);
-			};
+			return [
+				(a, b) => {
+					const { castA, castB } = applyCastFunctions(a, b);
+					return createAtomicValue(
+						addDurationToDateTime(castA.value, castB.value),
+						ValueType.XSDATE
+					);
+				},
+				ValueType.XSDATE,
+			];
 		}
 		if (operator === 'subtractOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					subtractDurationFromDateTime(castA.value, castB.value),
-					ValueType.XSDATE
-				);
-			};
+			return [
+				(a, b) => {
+					const { castA, castB } = applyCastFunctions(a, b);
+					return createAtomicValue(
+						subtractDurationFromDateTime(castA.value, castB.value),
+						ValueType.XSDATE
+					);
+				},
+				ValueType.XSDATE,
+			];
 		}
 	}
 
 	if (isSubtypeOf(typeA, ValueType.XSTIME) && isSubtypeOf(typeB, ValueType.XSDAYTIMEDURATION)) {
 		if (operator === 'addOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					addDurationToDateTime(castA.value, castB.value),
-					ValueType.XSTIME
-				);
-			};
+			return [
+				(a, b) => {
+					const { castA, castB } = applyCastFunctions(a, b);
+					return createAtomicValue(
+						addDurationToDateTime(castA.value, castB.value),
+						ValueType.XSTIME
+					);
+				},
+				ValueType.XSTIME,
+			];
 		}
 		if (operator === 'subtractOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					subtractDurationFromDateTime(castA.value, castB.value),
-					ValueType.XSTIME
-				);
-			};
+			return [
+				(a, b) => {
+					const { castA, castB } = applyCastFunctions(a, b);
+					return createAtomicValue(
+						subtractDurationFromDateTime(castA.value, castB.value),
+						ValueType.XSTIME
+					);
+				},
+				ValueType.XSTIME,
+			];
 		}
 	}
 
@@ -369,22 +445,28 @@ function generateBinaryOperatorFunction(
 			isSubtypeOf(typeA, ValueType.XSDATETIME))
 	) {
 		if (operator === 'addOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					addDurationToDateTime(castB.value, castA.value),
-					ValueType.XSDATETIME
-				);
-			};
+			return [
+				(a, b) => {
+					const { castA, castB } = applyCastFunctions(a, b);
+					return createAtomicValue(
+						addDurationToDateTime(castB.value, castA.value),
+						ValueType.XSDATETIME
+					);
+				},
+				ValueType.XSDATETIME,
+			];
 		}
 		if (operator === 'subtractOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					subtractDurationFromDateTime(castB.value, castA.value),
-					ValueType.XSDATETIME
-				);
-			};
+			return [
+				(a, b) => {
+					const { castA, castB } = applyCastFunctions(a, b);
+					return createAtomicValue(
+						subtractDurationFromDateTime(castB.value, castA.value),
+						ValueType.XSDATETIME
+					);
+				},
+				ValueType.XSDATETIME,
+			];
 		}
 	}
 
@@ -393,50 +475,83 @@ function generateBinaryOperatorFunction(
 		(isSubtypeOf(typeB, ValueType.XSYEARMONTHDURATION) && isSubtypeOf(typeA, ValueType.XSDATE))
 	) {
 		if (operator === 'addOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					addDurationToDateTime(castB.value, castA.value),
-					ValueType.XSDATE
-				);
-			};
+			return [
+				(a, b) => {
+					const { castA, castB } = applyCastFunctions(a, b);
+					return createAtomicValue(
+						addDurationToDateTime(castB.value, castA.value),
+						ValueType.XSDATE
+					);
+				},
+				ValueType.XSDATE,
+			];
 		}
 		if (operator === 'subtractOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					subtractDurationFromDateTime(castB.value, castA.value),
-					ValueType.XSDATE
-				);
-			};
+			return [
+				(a, b) => {
+					const { castA, castB } = applyCastFunctions(a, b);
+					return createAtomicValue(
+						subtractDurationFromDateTime(castB.value, castA.value),
+						ValueType.XSDATE
+					);
+				},
+				ValueType.XSDATE,
+			];
 		}
 	}
 
 	if (isSubtypeOf(typeB, ValueType.XSDAYTIMEDURATION) && isSubtypeOf(typeA, ValueType.XSTIME)) {
 		if (operator === 'addOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					addDurationToDateTime(castB.value, castA.value),
-					ValueType.XSTIME
-				);
-			};
+			return [
+				(a, b) => {
+					const { castA, castB } = applyCastFunctions(a, b);
+					return createAtomicValue(
+						addDurationToDateTime(castB.value, castA.value),
+						ValueType.XSTIME
+					);
+				},
+				ValueType.XSTIME,
+			];
 		}
 		if (operator === 'subtractOp') {
-			return (a, b) => {
-				const { castA, castB } = applyCastFunctions(a, b);
-				return createAtomicValue(
-					subtractDurationFromDateTime(castB.value, castA.value),
-					ValueType.XSTIME
-				);
-			};
+			return [
+				(a, b) => {
+					const { castA, castB } = applyCastFunctions(a, b);
+					return createAtomicValue(
+						subtractDurationFromDateTime(castB.value, castA.value),
+						ValueType.XSTIME
+					);
+				},
+				ValueType.XSTIME,
+			];
 		}
 	}
 
 	throw new Error(`XPTY0004: ${operator} not available for types ${typeA} and ${typeB}`);
 }
 
-const operatorsByTypingKey = Object.create(null);
+const operatorsByTypingKey: Record<string, [BinaryEvaluationFunction, ValueType]> = Object.create(
+	null
+);
+
+export function getBinaryPrefabOperator(
+	leftType: ValueType,
+	rightType: ValueType,
+	operator: string
+) {
+	const typingKey = `${leftType}~${rightType}~${operator}`;
+
+	let prefabOperator = operatorsByTypingKey[typingKey];
+	if (!prefabOperator) {
+		prefabOperator = operatorsByTypingKey[typingKey] = generateBinaryOperatorFunction(
+			operator,
+			leftType,
+			rightType
+		);
+	}
+
+	return prefabOperator;
+}
 
 class BinaryOperator extends Expression {
 	private _evaluateFunction?: BinaryEvaluationFunction;
@@ -510,19 +625,11 @@ class BinaryOperator extends Expression {
 					);
 				}
 
-				// TODO: Find a more errorproof solution, hash collisions can occur here
-				const typingKey = `${firstValue.type}~${secondValue.type}~${this._operator}`;
-				let prefabOperator = operatorsByTypingKey[typingKey];
-				if (!prefabOperator) {
-					prefabOperator = operatorsByTypingKey[
-						typingKey
-					] = generateBinaryOperatorFunction(
-						this._operator,
-						firstValue.type,
-						secondValue.type,
-						this.type
-					);
-				}
+				const prefabOperator = getBinaryPrefabOperator(
+					firstValue.type,
+					secondValue.type,
+					this._operator
+				)[0];
 
 				return sequenceFactory.singleton(prefabOperator(firstValue, secondValue));
 			});
